Rename JabRefPreferencesFilterDialog -> PreferencesFilterDialog and move it to gui
